+++ changelog.md
@@ -1,3 +1,4 @@
+- Fixed null pointer exception in Lidar Array AI. 
 - Improved Lidar Array AI; now the entire ship rotates to properly lead target with hardpoint weapons.
 
 Version 1.5.0